app/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
Jinja2==3.1.3
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
